Update patch set 2
Patch Set 2:

Will this work with cliff-2.6.0?, note that branch rpm-master is used to build both pike (which builds cliff from master head), and pike-uc (which currently builds cliff-2.6.0 https://github.com/redhat-openstack/rdoinfo/blob/master/rdo.yml#L2092).

If this change is not compatible with 2.6.0, we'll fix build in pike but will break it in pike-uc which is more important. In the worst case we can leave this patch unmerged until the commit which switched to testrepository is tagged and added in rdoinfo.

Patch-set: 2
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0